Split things out into sub classes for java's sake
